+++ src/kpuibase.h
@@ -132,7 +132,7 @@ public:
                                  bool reInitialize = false);
     virtual void SetSoundVolume(int volume) const;
     virtual void SetMusicVolume(int volume) const;
-    virtual void PlayAudio(int soundId) const;
+    virtual void PlayAudio(size_t soundId) const;
     virtual void PlayMusic(bool On, bool resetPos = false);
     virtual void StopMusicCallback();
